fixed autonomous difficulties, attempted modifications to drive PIDs
we got an auto play working, the errors we encountered were just formatting the auto play txt file incorrectly, and not fixing the file permission errors (I think I made a solution for this, but we'll see). actually another error was not passing through the control state in drive_set during autonomous, which would overwrite the control state and wouldn't change the control state to the command specified

for the drive_execute, we started adding some modifications. if the error of the angle of the azimuth motors is greater than a certain value, then the drive motors would not run as a result, to minimize the error when we start (which now that I think about it, might worsen our problem of overcoming static friction). we also added an aggressive/normal PID paradigm, where if the error is over a certain angle, the PID is shifted to a more aggressive form of it that moves it with more force, in order to reduce the error more quickly. I have no idea if these actually help or hurt, they were just ideas to try out.