changed parameters in M14 clusters and in execute VIs to be in arrays
in the flywheel, intake, and turret subsystems, which were imported in from M14, the parameters are individual values in the specific clusters. I think that it would be more suitable to have parameters in an array in order to make subsystem clusters more similar to each other. this is done so that in autonomous_load, the different cases in the case structure are much more similar to each other, since parameters aren't needed to be manually rewired for each subsystem depending on the different number of parameters. this way there is both more variability allowed, and less headache to set up the autonomous systems. 

additionally, before this, all of the subsystem's plays would have to have the same number of parameters, or have placeholder values, which never sat right for me. this way, the number of parameters can change depending on each play, because the last value in the passed in array will always be the concurrent value, and the rest are automatically passed through.