changed intake pneumatic value, added prevention of unintended movement
the intake begin value was set to 0 and 2, when it was 0 and 1

the azimuth motors would move wildly during No-Op, since the PID loops would still be active (could change this), so added a multiplier to remove azimuth motor outputs if in No-Op. We still experienced a strange behavior where no one would touch the controller, but every few seconds the control state of the drive would change to Manual_Control for a single loop iteration, then change back. I added a VI that averages the boolean value over a certain number of loop iterations, and so that won't happen anymore. It will add a slight delay to the beginning of No-Op changing to Manual_Control, but it shouldn't be much. I can change it if it's too noticeable, it only averages across 3 values so that should be a short amount of time.

we did have some very strange issues with deploying on this commit (LabVIEW would try to connect to target for many minutes, and would be softlocked if one tried to cancel it), so if the issues continue after this commit, there was something changed in one of these files that made it happen