changed azimuth sensors, tuned flywheel
we were using the relative sensor on the corner swerve sensors, when actually the relative encoder in the azimuth falcons is much faster and more reliable. we changed the angle sensor values to use these falcon encoders instead to prevent lag (especially in built code). the gear ratio from the falcons to the shaft was 12.8:1, but the falcon reads 2048 ticks per revolution, and the absolute sensor does 4096 per revolution, so we divided the absolute sensor by 2. we also added a way to zero the IMU, and now the field-centric works

we tuned the flywheel in phoenix tuner, we converted the RPM goal to ticks per 100ms, which let us use a low F value in the PIDF. the values in pheonix tuner are also written in flywheel_execute for future reference.

we also re-added the magazine delay to not run in fire until the flywheel is spun up