path planning, fixed joystick discrepancy between our code and 4265's
their positive X is forward, and their positive Y is left, ours was positive Y is forward and positive X is right, so we flipped the values in many places to make our code consistent with theirs, past the joystick calls in drive_teleop, positive X is forward, positive Y is left, and positive rotation is counterclockwise

so far the feed forward works for the path planning (after 5 hours), will require much more tuning, but its a start

the paths and plays in the auto folder is consistent with the ones on the robot, the latest is flipped from the others, since I didn't know which way the field was oriented relative to the physical field. in the 4265 path planner, the left side is the driver station side of the building, and vice versa

there's probably a bunch of other things, but most were related to path planning. i don't remember them though, oops