fixed boolean logic for end condition, fixed velocity conversion math
now the robot will now stop at the end of the play, though there is no distance from start calculation involved in the boolean logic since it didn't seem to make much sense to have it

changed the strafe maximum in joystickstrafescaling from 0.5 to 1 with a value passed in, since before it would go much slower than we wanted it to go

wrote a lot more test auto plays and plans, currently the plan implemented is MARS-PathPlanningTestWithIntakingAndTwoSteps.txt (which probably should change name), which does two paths, MARS-Heading-Tuning-Test2, and MARS-Heading-Tuning-Test-Part-2. currently there isn't a way to pass in a string parameter, so i had to do a dumb workaround in drive_execute to get it to work for the video. i'll probably change the parameters for the drive cluster from an array to a cluster so it can bring strings and numerics in, we'll see if it works